requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Migrate==4.0.5
mysql-connector-python==8.2.0
werkzeug==3.0.1
email-validator==2.1.0.post1
openai==1.3.0
google-generativeai==0.3.1
python-dotenv==1.0.0
reportlab==4.1.0
msal==1.28.0
